fix Playlist.OrderedList() breaking the index of all songs
Since 26f0033ebb5f8e7c204a042709e3dfa885eac868 and using Song pointers
the copied Playlist in OrderedList actually holds the same
Songs as the originally one.
This and the side effects in Playlist.Pop (setting the index to -1)
breaks the heap because all songs in the original queue now
have a invalid index (-1).

This is fixed by removing the side effect in Playlist.Pop().